requirements.txt
fastapi
uvicorn
azure-ai-formrecognizer
azure-core
pydantic
python-dotenv
